Refactor type test generation code
Move the type test generation code to a separate class to not increase the size of the JvmInstructionGen class.
